+++ src/ssl.rs
@@ -2,7 +2,7 @@ use std::{fmt, ptr};
 
 use open62541_sys::UA_CreateCertificate;
 
-use crate::{ua, DataType, Error, Result};
+use crate::{ua, DataType, Error};
 
 /// Certificate in [DER] or [PEM] format.
 ///
@@ -34,6 +34,26 @@ impl Certificate {
     pub fn as_bytes(&self) -> &[u8] {
         // SAFETY: We always initialize inner value.
         unsafe { self.0.as_bytes_unchecked() }
+    }
+
+    /// Parses certificate.
+    ///
+    /// # Errors
+    ///
+    /// This fails when the certificate cannot be parsed or is invalid.
+    #[cfg(feature = "x509")]
+    pub fn into_x509(
+        self,
+    ) -> Result<x509_certificate::X509Certificate, x509_certificate::X509CertificateError> {
+        use x509_certificate::{X509Certificate, X509CertificateError};
+
+        // Apply heuristic to get certificates from both DER and PEM format. Try PEM first because
+        // the implementation first extracts DER data from PEM and can tell us whether this failed
+        // (or the certificate itself was invalid).
+        X509Certificate::from_pem(self.as_bytes()).or_else(|err| match err {
+            X509CertificateError::PemDecode(_) => X509Certificate::from_der(self.as_bytes()),
+            err => Err(err),
+        })
     }
 
     pub(crate) const fn as_byte_string(&self) -> &ua::ByteString {
@@ -117,7 +137,7 @@ pub fn create_certificate(
     subject_alt_name: &ua::Array<ua::String>,
     cert_format: &ua::CertificateFormat,
     params: Option<&ua::KeyValueMap>,
-) -> Result<(Certificate, PrivateKey)> {
+) -> crate::Result<(Certificate, PrivateKey)> {
     // Create logger that forwards to Rust `log`. It is only used for the function call below and it
     // will be cleaned up at the end of the function.
     let mut logger = ua::Logger::rust_log();